+++ example/lib/main.dart
@@ -89,9 +89,30 @@ class _MyHomePageState extends State<MyHomePage> {
   Future<Null> _cropImage() async {
     File croppedFile = await ImageCropper.cropImage(
       sourcePath: imageFile.path,
-      toolbarTitle: 'Cropper',
-      toolbarColor: Colors.blue,
-      toolbarWidgetColor: Colors.white,
+      aspectRatioPresets: Platform.isAndroid
+          ? [
+              CropAspectRatioPreset.square,
+              CropAspectRatioPreset.ratio3x2,
+              CropAspectRatioPreset.original,
+              CropAspectRatioPreset.ratio4x3,
+              CropAspectRatioPreset.ratio16x9
+            ]
+          : [
+              CropAspectRatioPreset.original,
+              CropAspectRatioPreset.square,
+              CropAspectRatioPreset.ratio3x2,
+              CropAspectRatioPreset.ratio4x3,
+              CropAspectRatioPreset.ratio5x3,
+              CropAspectRatioPreset.ratio5x4,
+              CropAspectRatioPreset.ratio7x5,
+              CropAspectRatioPreset.ratio16x9
+            ],
+      androidUiSettings: AndroidUiSettings(
+          toolbarTitle: 'Cropper',
+          toolbarColor: Colors.deepOrange,
+          toolbarWidgetColor: Colors.white,
+          initAspectRatio: CropAspectRatioPreset.original,
+          lockAspectRatio: false),
     );
     if (croppedFile != null) {
       imageFile = croppedFile;
@@ -99,10 +120,6 @@ class _MyHomePageState extends State<MyHomePage> {
         state = AppState.cropped;
       });
     }
-  }
-
-  Future<Null> _saveImage() async {
-
   }
 
   void _clearImage() {